requirements.txt
bleach==2.0.0
click==6.7
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.2
Flask==0.12.2
future==0.16.0
html5lib==0.999999999
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mpmath==0.19
nbconvert==5.1.1
nbformat==4.3.0
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numpy==1.12.1
oauthlib==2.0.2
pandas==0.20.1
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-twitter==3.3
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.14.2
requests-oauthlib==0.8.0
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
sympy==1.0
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
tweepy==3.5.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
